[CHORE] #59 - move signin, signup activity to user package
