+++ app/src/main/AndroidManifest.xml
@@ -32,18 +32,18 @@
             android:name=".ui.view.daily.detail.DailyDetailActivity"
             android:windowSoftInputMode="adjustPan" />
         <activity
-            android:name=".ui.view.signin.SignInActivity"
+            android:name=".ui.view.user.SignInActivity"
             android:exported="true" />
         <activity
             android:name=".ui.view.goal.detail.GoalDetailActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustResize" />
         <activity
-            android:name=".ui.view.signup.SignUpActivity"
+            android:name=".ui.view.user.SignUpActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustPan" />
         <activity
-            android:name=".ui.view.signup.SignUpTermsActivity"
+            android:name=".ui.view.user.SignUpTermsActivity"
             android:exported="true" />
         <activity android:name=".ui.view.MainActivity">
             <intent-filter>
